requirements.txt
PyMuPDF==1.24.5
     spacy==3.7.2
     sentence-transformers==2.7.0
     faiss-cpu==1.8.0
     xgboost==2.0.3
     ollama==0.1.9
     streamlit==1.36.0
     pyyaml==6.0.1
     numpy==1.26.4
     pandas==2.2.2
     scikit-learn==1.4.2
     nltk==3.8.1